+++ src/customEmojis.js
@@ -1,0 +1,9 @@
+const donatorEmoji = '<:blobcatheart:873846489292619776>';
+const developerEmoji = '<:blob_dev:891643826895212575>';
+const senseiEmoji = '<:verified:891645797874151445>';
+
+module.exports = {
+	donatorEmoji,
+	developerEmoji,
+	senseiEmoji,
+};